requirements.txt
Pillow
typeguard==2.2.2
nibabel==2.3.0
scipy==1.1.0
torch==1.7.1
torchsummay==1.5.1
pytorch-msssim